Fix bug when crash fails template may never stop.
Template works by replacing "> {{var}}" to an anonymous pipe,
then passing the modified command to crash shell to execute it.
In the mean time, template will cat a named pipe and wait for response
from crash. Normally it works fine, but if crash fails in the middle,
the process catting named pipe will never stop. This patch deals with
the problem.

Template will now continuously echo \n to the anonymous pipe which passes
the modify command to crash, if crash fails, template will receive
SIGPIPE and exit. Thus avoid stopless.

Signed-off-by: Tao Liu <ltao@redhat.com>