Remove temporary file dependency of content compare
Previously, if we want to compare the content of 2 virtual
addresses, command file will redirect the contents into 2
temporarty files, then compare the contents of the file. It
will fail if we run in multi-thread(-u), because the 2 files
names are identical among multi-thread.

This patch removes the temporary file dependency, thus multi-
thread safe.

Signed-off-by: Tao Liu <ltao@redhat.com>