Fix when multi-thread running the template sometimes will timeout and hangup
Sometimes in multi-thread, one crash thread will take long time to
schedule, previously the template gives commands to crash, starts time
counting down, and waits for value sent back. It doesn't exclude the
re-scheduling time. So the real time given to crash is usually shorter
than expected.

Besides, when template timeout it will exit. But when crash re-scheduled
and trying to write to named pipe to send value back, there is no pipe
reader, thus crash will wait and hang.

This patch first add a function to sync template and crash, ensure crash
is responsive then start time counting down. If template really timeout,
it will delete the named pipe. So if crash re-scheduled and redirect
output to a file, it will be writing to a regular file with the same name.

Signed-off-by: Tao Liu <ltao@redhat.com>