Remove cpu arch dependent hook
Previously, a cpu arch dependent hook was introduced, because we
didn't have a good log filter to identify an error when crash fails.

Currently we have log_filter.sh, in which a batch of error featured
strings are introduced. Thus depending on injecting error processing
functions within crash to identify a crash failing is not as important
as before when comparing with the extra effort of maintaining the
cpu arch dependent hook.

This patch removes the hook, making the whole project cpu arch independent.

Signed-off-by: Tao Liu <ltao@redhat.com>