Fix bug getting exit value error
Before commit "Implemented a feature string based log filter",
the loop getting line of dumplist file is through:

  cat xxx_file | while read ARG1; do
     xxxxx
  done

After the commit, it becomes:

  while read ARG1; do
     xxxxx
  done <<< $(cat xxx_file)

The difference is the former xxxxx running in another process, the
latter running in current process. Thus exit as xxxxx in the latter
will directly quit the current process, which is not we expected.

This patch changed the way to exit the loop.

Signed-off-by: Tao Liu <ltao@redhat.com>