Reimplemented the log_filter
This patch reimplemented the regex based log_filter for specific test
cases, and pretty the output format to reduce the output log.

Previously we used a generic log_filter for all test commands, which
could identify and deny error keywords, but it was lack of agility. An
error keywords should be denied for one test command, but may be allowed
for another. So we introduced the regex based log_filter, which varied
regex rules case by case.

Signed-off-by: Tao Liu <ltao@redhat.com>