Fix bug when live testing, log_filter will have to filter itself
When live testing, the testing scripts will be part of ENV for specific
processes. Some crash testing commands as "ps -a" will output ENV for
all processes, thus the scripts themselves are outputed as part of the log.

Because log_filter contains a lot of error identification keywords, thus
it will be difficult to handle the log_filter filter itself case.

In this patch, we simply see if it is live testing, then deny all outputs
of "ps -a", to bypass the situation.

Signed-off-by: Tao Liu <ltao@redhat.com>