Add dumplist_generator.sh to automatically create dumplist
Previously, dumplist file is created by hand. If user have a large collection
of vmcores, it will be inconvenient to create one.

This patch indroduced a dumplist_generator, which can scan the directory and
subdirectories which holding vmlinux and vmcore, and automatically generate
the dumplist file for later use of crash_test.sh.

Signed-off-by: Tao Liu <ltao@redhat.com>